requirements.txt
# Web scraping and automated browser
selenium
webdriver-manager
requests
beautifulsoup4

# Database
psycopg2-binary

# Scheduling and recurring tasks
apscheduler
python-dotenv

# OpenAI API client
openai
azure-ai-ml

# Console processes
tqdm